Remove unwanted cast of DO_GET_REFCNT to long int in qbuffer.c [Test-built runtime on linux32 (cflbs32) and linux64 (chap02)]
git-svn-id: http://svn.code.sf.net/p/chapel/code/trunk@22168 3a8e244f-b0f2-452b-bcba-4c88e055c3ca

+++ runtime/src/qio/qbuffer.c
@@ -74,7 +74,7 @@ void qbytes_free_iobuf(qbytes_t* b) {
 void debug_print_bytes(qbytes_t* b)
 {
   fprintf(stderr, "bytes %p: data=%p len=%lli ref_cnt=%" PRIu64 " free_function=%p flags=%i\n",
-          b, b->data, (long long int) b->len, (long int) DO_GET_REFCNT(b),
+          b, b->data, (long long int) b->len, DO_GET_REFCNT(b),
           b->free_function, b->flags);
 }
